pyproject.toml
[project]
name = "python-genai-intro"
version = "0.1.0"
description = "Introduction to programming Gen-AI applications in Python"
authors = [{ name = "Matei Tene", email = "matei.tene@gmail.com" }]
requires-python = ">=3.12,<4.0.0"
readme = "README.md"
dependencies = [
    "streamlit>=1.54.0,<1.55.0",
    "streamlit-chat>=0.1.1,<0.2.0",
    "langchain>=1.2.10,<1.3.0",
    "langchain-chroma>=1.1.0,<1.2.0",
    "langchain-ollama>=1.0.1,<1.1.0",
    "langchain-text-splitters>=1.1.0,<1.2.0",
    "langchain-mcp-adapters>=0.2.1,<0.3.0",
    "langchain-openai>=1.1.6,<1.2.0",
    "deepagents>=0.4.5,<0.5.0",
    "langgraph>=1.0.10,<1.1.0",
    "a2a-sdk>=0.3.24,<0.4.0",
    "fastmcp>=3.1.0,<3.2.0",
    "pydantic>=2.12.5,<2.13.0",
    "pyyaml>=6.0.3,<6.1.0",
    "rich>=14.1.0,<14.2.0",
    "tzdata>=2025.3,<2026.0",
    "opentelemetry-sdk>=1.39.1,<1.40.0",
    "opentelemetry-exporter-otlp>=1.39.1,<1.40.0",
    "opentelemetry-instrumentation-langchain>=0.51.1,<0.52.0",
    "opentelemetry-instrumentation-openai-v2>=2.3b0,<2.4.0",
    "opentelemetry-instrumentation-chromadb>=0.52.2,<0.53.0",
    "opentelemetry-instrumentation-requests>=0.60b1,<0.70",
    "opentelemetry-instrumentation-httpx>=0.60b1,<0.70",
]

[dependency-groups]
dev = [
    "ruff~=0.14.0",
    "pyright~=1.1.0",
]

[project.scripts]
genai-chat = "user_interface.__main__:main"
exercise-0 = "chatbot.lessons.exercises.e00_intro.__main__:main"
exercise-1 = "chatbot.lessons.exercises.e01_prompting.__main__:main"
solution-1 = "chatbot.lessons.solutions.s01_prompting.__main__:main"
exercise-2 = "chatbot.lessons.exercises.e02_system_prompt.__main__:main"
solution-2 = "chatbot.lessons.solutions.s02_system_prompt.__main__:main"
exercise-3 = "chatbot.lessons.exercises.e03_memory.__main__:main"
solution-3 = "chatbot.lessons.solutions.s03_memory.__main__:main"
exercise-4 = "chatbot.lessons.exercises.e04_rag.__main__:main"
solution-4 = "chatbot.lessons.solutions.s04_rag.__main__:main"
exercise-5 = "chatbot.lessons.exercises.e05_structured_outputs.__main__:main"
solution-5 = "chatbot.lessons.solutions.s05_structured_outputs.__main__:main"
exercise-6 = "chatbot.lessons.exercises.e06_tool_calling.__main__:main"
solution-6 = "chatbot.lessons.solutions.s06_tool_calling.__main__:main"
exercise-7 = "chatbot.lessons.exercises.e07_mcp.__main__:main"
solution-7 = "chatbot.lessons.solutions.s07_mcp.__main__:main"
exercise-8 = "chatbot.lessons.exercises.e08_custom_agent.__main__:main"
solution-8 = "chatbot.lessons.solutions.s08_custom_agent.__main__:main"
exercise-9 = "chatbot.lessons.exercises.e09_deep_agent.__main__:main"
solution-9 = "chatbot.lessons.solutions.s09_deep_agent.__main__:main"
exercise-10 = "chatbot.lessons.exercises.e10_a2a.__main__:main"
solution-10 = "chatbot.lessons.solutions.s10_a2a.__main__:main"

[tool.hatch.build.targets.sdist]
include = [
    "src/user_interface",
    "src/chatbot",
]

[tool.hatch.build.targets.wheel]
include = [
    "src/user_interface",
    "src/chatbot",
]

[tool.hatch.build.targets.wheel.sources]
"src/user_interface" = "user_interface"
"src/chatbot" = "chatbot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
